+++ src/test/java/edu/illinois/library/cantaloupe/resource/ImageRequestHandlerTest.java
@@ -1,5 +1,6 @@
 package edu.illinois.library.cantaloupe.resource;
 
+import edu.illinois.library.cantaloupe.Application;
 import edu.illinois.library.cantaloupe.cache.CacheFacade;
 import edu.illinois.library.cantaloupe.cache.CompletableOutputStream;
 import edu.illinois.library.cantaloupe.cache.DerivativeCache;
@@ -11,29 +12,34 @@ import edu.illinois.library.cantaloupe.operation.OperationList;
 import edu.illinois.library.cantaloupe.operation.ValidationException;
 import edu.illinois.library.cantaloupe.processor.Processor;
 import edu.illinois.library.cantaloupe.delegate.DelegateProxy;
+import edu.illinois.library.cantaloupe.processor.SourceFormatException;
 import edu.illinois.library.cantaloupe.test.BaseTest;
 import edu.illinois.library.cantaloupe.test.TestUtil;
+import edu.illinois.library.cantaloupe.test.WebServer;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 
 import java.io.ByteArrayOutputStream;
 import java.io.OutputStream;
+import java.nio.charset.StandardCharsets;
+import java.nio.file.Files;
+import java.nio.file.Path;
 
 import static org.junit.jupiter.api.Assertions.*;
 
-public class ImageRequestHandlerTest extends BaseTest {
+class ImageRequestHandlerTest extends BaseTest {
 
     @Nested
-    public class BuilderTest {
+    class BuilderTest {
 
         @Test
-        void testBuildWithNoOperationListSet() {
+        void buildWithNoOperationListSet() {
             assertThrows(NullPointerException.class, () ->
                     ImageRequestHandler.builder().build());
         }
 
         @Test
-        void testOptionallyWithDelegateProxyWithNonNullArguments() {
+        void optionallyWithDelegateProxyWithNonNullArguments() {
             DelegateProxy proxy = TestUtil.newDelegateProxy();
             ImageRequestHandler handler = ImageRequestHandler.builder()
                     .optionallyWithDelegateProxy(proxy, proxy.getRequestContext())
@@ -44,7 +50,7 @@ public class ImageRequestHandlerTest extends BaseTest {
         }
 
         @Test
-        void testOptionallyWithDelegateProxyWithNullDelegateProxy() {
+        void optionallyWithDelegateProxyWithNullDelegateProxy() {
             RequestContext context = new RequestContext();
             ImageRequestHandler handler = ImageRequestHandler.builder()
                     .optionallyWithDelegateProxy(null, context)
@@ -56,7 +62,7 @@ public class ImageRequestHandlerTest extends BaseTest {
         }
 
         @Test
-        void testOptionallyWithDelegateProxyWithNullRequestContext() {
+        void optionallyWithDelegateProxyWithNullRequestContext() {
             DelegateProxy delegateProxy = TestUtil.newDelegateProxy();
             ImageRequestHandler handler = ImageRequestHandler.builder()
                     .withOperationList(new OperationList())
@@ -68,7 +74,7 @@ public class ImageRequestHandlerTest extends BaseTest {
         }
 
         @Test
-        void testOptionallyWithDelegateProxyWithNullArguments() {
+        void optionallyWithDelegateProxyWithNullArguments() {
             ImageRequestHandler handler = ImageRequestHandler.builder()
                     .withOperationList(new OperationList())
                     .optionallyWithDelegateProxy(null, null)
@@ -79,14 +85,14 @@ public class ImageRequestHandlerTest extends BaseTest {
         }
 
         @Test
-        void testWithDelegateProxyWithNullDelegateProxy() {
+        void withDelegateProxyWithNullDelegateProxy() {
             assertThrows(IllegalArgumentException.class, () ->
                     ImageRequestHandler.builder()
                             .withDelegateProxy(null, new RequestContext()));
         }
 
         @Test
-        void testWithDelegateProxyWithNullRequestContext() {
+        void withDelegateProxyWithNullRequestContext() {
             DelegateProxy delegateProxy = TestUtil.newDelegateProxy();
             assertThrows(IllegalArgumentException.class, () ->
                     ImageRequestHandler.builder()
@@ -129,13 +135,13 @@ public class ImageRequestHandlerTest extends BaseTest {
     }
 
     @Test
-    void testHandleCallsPreAuthorizationCallback() throws Exception {
+    void handleCallsPreAuthorizationCallback() throws Exception {
         {   // Configure the application.
             final Configuration config = Configuration.getInstance();
             config.setProperty(Key.CACHE_SERVER_RESOLVE_FIRST, false);
             config.setProperty(Key.SOURCE_STATIC, "FilesystemSource");
             config.setProperty(Key.FILESYSTEMSOURCE_PATH_PREFIX,
-                    TestUtil.getImagesPath().toString() + "/");
+                    TestUtil.getImagesPath() + "/");
         }
 
         // Configure the request.
@@ -155,13 +161,13 @@ public class ImageRequestHandlerTest extends BaseTest {
     }
 
     @Test
-    void testHandleCallsAuthorizationCallback() throws Exception {
+    void handleCallsAuthorizationCallback() throws Exception {
         {   // Configure the application.
             final Configuration config = Configuration.getInstance();
             config.setProperty(Key.CACHE_SERVER_RESOLVE_FIRST, false);
             config.setProperty(Key.SOURCE_STATIC, "FilesystemSource");
             config.setProperty(Key.FILESYSTEMSOURCE_PATH_PREFIX,
-                    TestUtil.getImagesPath().toString() + "/");
+                    TestUtil.getImagesPath() + "/");
         }
 
         // Configure the request.
@@ -181,13 +187,13 @@ public class ImageRequestHandlerTest extends BaseTest {
     }
 
     @Test
-    void testHandleCallsCacheStreamingCallback() throws Exception {
+    void handleCallsCacheStreamingCallback() throws Exception {
         {   // Configure the application.
             final Configuration config = Configuration.getInstance();
             config.setProperty(Key.CACHE_SERVER_RESOLVE_FIRST, false);
             config.setProperty(Key.SOURCE_STATIC, "FilesystemSource");
             config.setProperty(Key.FILESYSTEMSOURCE_PATH_PREFIX,
-                    TestUtil.getImagesPath().toString() + "/");
+                    TestUtil.getImagesPath() + "/");
             config.setProperty(Key.DERIVATIVE_CACHE_ENABLED, true);
             config.setProperty(Key.DERIVATIVE_CACHE, "HeapCache");
         }
@@ -233,13 +239,13 @@ public class ImageRequestHandlerTest extends BaseTest {
     }
 
     @Test
-    void testHandleCallsInfoAvailableCallback() throws Exception {
+    void handleCallsInfoAvailableCallback() throws Exception {
         { // Configure the application.
             final Configuration config = Configuration.getInstance();
             config.setProperty(Key.CACHE_SERVER_RESOLVE_FIRST, false);
             config.setProperty(Key.SOURCE_STATIC, "FilesystemSource");
             config.setProperty(Key.FILESYSTEMSOURCE_PATH_PREFIX,
-                    TestUtil.getImagesPath().toString() + "/");
+                    TestUtil.getImagesPath() + "/");
         }
 
         // Configure the request.
@@ -259,13 +265,13 @@ public class ImageRequestHandlerTest extends BaseTest {
     }
 
     @Test
-    void testHandleCallsProcessingCallback() throws Exception {
+    void handleCallsProcessingCallback() throws Exception {
         { // Configure the application.
             final Configuration config = Configuration.getInstance();
             config.setProperty(Key.CACHE_SERVER_RESOLVE_FIRST, false);
             config.setProperty(Key.SOURCE_STATIC, "FilesystemSource");
             config.setProperty(Key.FILESYSTEMSOURCE_PATH_PREFIX,
-                    TestUtil.getImagesPath().toString() + "/");
+                    TestUtil.getImagesPath() + "/");
         }
 
         // Configure the request.
@@ -285,12 +291,12 @@ public class ImageRequestHandlerTest extends BaseTest {
     }
 
     @Test
-    void testHandleProcessesImage() throws Exception {
+    void handleProcessesImage() throws Exception {
         { // Configure the application.
             final Configuration config = Configuration.getInstance();
             config.setProperty(Key.SOURCE_STATIC, "FilesystemSource");
             config.setProperty(Key.FILESYSTEMSOURCE_PATH_PREFIX,
-                    TestUtil.getImagesPath().toString() + "/");
+                    TestUtil.getImagesPath() + "/");
         }
 
         // Configure the request.
@@ -310,13 +316,13 @@ public class ImageRequestHandlerTest extends BaseTest {
     }
 
     @Test
-    void testHandleStreamsFromDerivativeCache() throws Exception {
+    void handleStreamsFromDerivativeCache() throws Exception {
         {   // Configure the application.
             final Configuration config = Configuration.getInstance();
             config.setProperty(Key.CACHE_SERVER_RESOLVE_FIRST, false);
             config.setProperty(Key.SOURCE_STATIC, "FilesystemSource");
             config.setProperty(Key.FILESYSTEMSOURCE_PATH_PREFIX,
-                    TestUtil.getImagesPath().toString() + "/");
+                    TestUtil.getImagesPath() + "/");
             config.setProperty(Key.DERIVATIVE_CACHE_ENABLED, true);
             config.setProperty(Key.DERIVATIVE_CACHE, "HeapCache");
         }
@@ -363,12 +369,12 @@ public class ImageRequestHandlerTest extends BaseTest {
     }
 
     @Test
-    void testHandleWithFailedPreAuthorization() throws Exception {
+    void handleWithFailedPreAuthorization() throws Exception {
         { // Configure the application.
             final Configuration config = Configuration.getInstance();
             config.setProperty(Key.SOURCE_STATIC, "FilesystemSource");
             config.setProperty(Key.FILESYSTEMSOURCE_PATH_PREFIX,
-                    TestUtil.getImagesPath().toString() + "/");
+                    TestUtil.getImagesPath() + "/");
         }
 
         // Configure the request.
@@ -405,12 +411,12 @@ public class ImageRequestHandlerTest extends BaseTest {
     }
 
     @Test
-    void testHandleWithFailedAuthorization() throws Exception {
+    void handleWithFailedAuthorization() throws Exception {
         { // Configure the application.
             final Configuration config = Configuration.getInstance();
             config.setProperty(Key.SOURCE_STATIC, "FilesystemSource");
             config.setProperty(Key.FILESYSTEMSOURCE_PATH_PREFIX,
-                    TestUtil.getImagesPath().toString() + "/");
+                    TestUtil.getImagesPath() + "/");
         }
 
         // Configure the request.
@@ -447,12 +453,12 @@ public class ImageRequestHandlerTest extends BaseTest {
     }
 
     @Test
-    void testHandleWithIllegalPageIndex() throws Exception {
+    void handleWithIllegalPageIndex() throws Exception {
         { // Configure the application.
             final Configuration config = Configuration.getInstance();
             config.setProperty(Key.SOURCE_STATIC, "FilesystemSource");
             config.setProperty(Key.FILESYSTEMSOURCE_PATH_PREFIX,
-                    TestUtil.getImagesPath().toString() + "/");
+                    TestUtil.getImagesPath() + "/");
         }
 
         // Configure the request.
@@ -474,12 +480,12 @@ public class ImageRequestHandlerTest extends BaseTest {
     }
 
     @Test
-    void testHandleWithInvalidOperationList() throws Exception {
+    void handleWithInvalidOperationList() throws Exception {
         { // Configure the application.
             final Configuration config = Configuration.getInstance();
             config.setProperty(Key.SOURCE_STATIC, "FilesystemSource");
             config.setProperty(Key.FILESYSTEMSOURCE_PATH_PREFIX,
-                    TestUtil.getImagesPath().toString() + "/");
+                    TestUtil.getImagesPath() + "/");
         }
 
         // Configure the request.
@@ -494,6 +500,134 @@ public class ImageRequestHandlerTest extends BaseTest {
              ByteArrayOutputStream outputStream = new ByteArrayOutputStream()) {
             assertThrows(ValidationException.class, () ->
                     handler.handle(outputStream));
+        }
+    }
+
+    @Test
+    void handleDeletesIncompatibleSourceCachedImageWhenSoConfigured()
+            throws Exception {
+        final WebServer server = new WebServer();
+        try {
+            server.start();
+
+            {   // Configure the application.
+                final Configuration config = Configuration.getInstance();
+                config.setProperty(Key.SOURCE_STATIC, "HttpSource");
+                config.setProperty(Key.HTTPSOURCE_URL_PREFIX,
+                        server.getHTTPURI().toString() + "/");
+                config.setProperty(Key.PROCESSOR_FALLBACK,
+                        edu.illinois.library.cantaloupe.processor.MockStreamProcessor.class.getName());
+                config.setProperty(Key.PROCESSOR_STREAM_RETRIEVAL_STRATEGY,
+                        "CacheStrategy");
+                config.setProperty(Key.PROCESSOR_PURGE_INCOMPATIBLE_FROM_SOURCE_CACHE,
+                        true); // what this test is testing
+                config.setProperty(Key.SOURCE_CACHE, "FilesystemCache");
+                config.setProperty(Key.SOURCE_CACHE_TTL, 300);
+                config.setProperty(Key.FILESYSTEMCACHE_PATHNAME,
+                        Application.getTempPath().toString());
+            }
+
+            // Configure the request.
+            final OperationList opList = new OperationList();
+            final Identifier identifier = new Identifier("jpg-rgb-64x48x8.jpg");
+            final Metadata metadata = new Metadata();
+            opList.setIdentifier(identifier);
+            Encode encode = new Encode(Format.get("jpg"));
+            encode.setCompression(Compression.JPEG);
+            encode.setQuality(80);
+            encode.setMetadata(metadata);
+            opList.add(encode);
+
+            final CacheFacade cacheFacade = new CacheFacade();
+
+            try (ImageRequestHandler handler = ImageRequestHandler.builder()
+                    .withOperationList(opList)
+                    .build();
+                 OutputStream outputStream = OutputStream.nullOutputStream()) {
+                // The first request should cause the source image to be
+                // source-cached...
+                handler.handle(outputStream);
+                // Overwrite the source-cached image with garbage, destroying
+                // any format-signifying magic bytes.
+                Path file = cacheFacade.getSourceCacheFile(identifier).get();
+                Files.write(file, "This is garbage".getBytes(StandardCharsets.UTF_8));
+                // Send the same request again. The source cache will be
+                // consulted instead of the source.
+                handler.handle(outputStream);
+                fail("Expected a SourceFormatException");
+            } catch (SourceFormatException e) {
+                // The delete happens asynchronously, so give it some time.
+                Thread.sleep(2000);
+                assertFalse(cacheFacade.getSourceCacheFile(identifier).isPresent());
+            }
+        } finally {
+            server.stop();
+        }
+    }
+
+    @Test
+    void handleDoesNotDeleteIncompatibleSourceCachedImageWhenNotConfiguredTo()
+            throws Exception {
+        final WebServer server = new WebServer();
+        try {
+            server.start();
+
+            {   // Configure the application.
+                final Configuration config = Configuration.getInstance();
+                config.setProperty(Key.SOURCE_STATIC, "HttpSource");
+                config.setProperty(Key.HTTPSOURCE_URL_PREFIX,
+                        server.getHTTPURI().toString() + "/");
+                config.setProperty(Key.PROCESSOR_FALLBACK,
+                        edu.illinois.library.cantaloupe.processor.MockStreamProcessor.class.getName());
+                config.setProperty(Key.PROCESSOR_STREAM_RETRIEVAL_STRATEGY,
+                        "CacheStrategy");
+                config.setProperty(Key.PROCESSOR_PURGE_INCOMPATIBLE_FROM_SOURCE_CACHE,
+                        false); // what this test is testing
+                config.setProperty(Key.SOURCE_CACHE, "FilesystemCache");
+                config.setProperty(Key.SOURCE_CACHE_TTL, 300);
+                config.setProperty(Key.FILESYSTEMCACHE_PATHNAME,
+                        Application.getTempPath().toString());
+            }
+
+            // Configure the request.
+            final OperationList opList = new OperationList();
+            final Identifier identifier = new Identifier("jpg-rgb-64x48x8.jpg");
+            final Metadata metadata = new Metadata();
+            opList.setIdentifier(identifier);
+            Encode encode = new Encode(Format.get("jpg"));
+            encode.setCompression(Compression.JPEG);
+            encode.setQuality(80);
+            encode.setMetadata(metadata);
+            opList.add(encode);
+
+            final CacheFacade cacheFacade = new CacheFacade();
+
+            try (ImageRequestHandler handler = ImageRequestHandler.builder()
+                    .withOperationList(opList)
+                    .build();
+                 OutputStream outputStream = OutputStream.nullOutputStream()) {
+                // The first request should cause the source image to be
+                // source-cached...
+                handler.handle(outputStream);
+                // Overwrite the source-cached image with garbage, destroying
+                // any format-signifying magic bytes.
+                Path file = cacheFacade.getSourceCacheFile(identifier).get();
+                Files.write(file, "This is garbage".getBytes(StandardCharsets.UTF_8));
+                // Send the same request again. The source cache will be
+                // consulted instead of the source.
+                handler.handle(outputStream);
+                fail("Expected a SourceFormatException");
+            } catch (SourceFormatException e) {
+                // The source-cached file is not supposed to get deleted, but
+                // it will happen asynchronously if it does, so give it some
+                // time.
+                Thread.sleep(2000);
+                Path file = cacheFacade.getSourceCacheFile(identifier).get();
+                assertTrue(Files.exists(file));
+                Files.delete(file);
+            }
+        } finally {
+            server.stop();
         }
     }