+++ CHANGES.md
@@ -2,11 +2,13 @@
 
 # 6.0
 
-* The health endpoint is en/disabled via `endpoint.health.enabled` rather than
+* The health endpoint is enabled via `endpoint.health.enabled` rather than
   `endpoint.api.enabled`.
 * HttpSource can be configured to send a ranged GET request instead of a HEAD
   request, enabling it to work with pre-signed URLs that do not allow HEAD
   requests.
+* Added a configuration option to automatically purge source-cached images
+  whose format cannot be inferred.
 * Added an HTTP API method to purge all infos from the derivative cache.
 * The delegate script pathname can be set using the
   `-Dcantaloupe.delegate_script` VM argument, which takes precedence over the